Refactored dmarc
